backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-solo==2.0.0
djangorestframework==3.14.0
numpy==1.23.4
opencv-contrib-python==4.6.0.66
Pillow==9.2.0
psycopg2-binary==2.9.5
pytz==2022.4
sqlparse==0.4.3
gunicorn==20.1.0
